Put the colons back.
Ok, I think I understand how this works.  The macro
DEFINE_ARTDAQ_COMMANDABLE_GENERATOR doesn't have anything
to do with how the name in a fcl is attached to a module.
That seems to be set in the CMakeLists.txt, and it currently
matches ("CRTFragGen") between there and the fcl.